summer_school_day4_vlms/requirements.txt
torch==2.7.1
huggingface_hub
ollama
Pillow
Requests
transformers
datasets
